Implement RedundantOverloadTypeParams
instead of UselessOverloadTypeParams
