+++ lib/rubocop/cop/rbs/lint/redundant_overload_type_params.rb
@@ -11,12 +11,15 @@ module RuboCop
         #   def foo: [T] () -> void
         #
         #   # bad
-        #   def bar: [T, U] (T) -> void
+        #   def bar: [T] () -> T
+        #
+        #   # bad
+        #   def baz: [T] () { () -> T } -> void
         #
         #   # good
         #   def foo: [T] (Array[T]) -> T
-        class UselessOverloadTypeParams < RuboCop::RBS::CopBase
-          MSG = 'Useless overload type variable - `%<variable>s`.'
+        class RedundantOverloadTypeParams < RuboCop::RBS::CopBase
+          MSG = 'Redundant overload type variable - `%<variable>s`.'
 
           def on_rbs_def(decl)
             decl.overloads.each do |overload|
@@ -24,7 +27,16 @@ module RuboCop
 
               type_params = overload.method_type.type_params
 
-              overload.method_type.each_type do |type|
+              types = []
+              overload.method_type.type.each_param do |param|
+                types << param.type
+              end
+              overload.method_type.block&.then do |block|
+                block.type.each_type do |t|
+                  types << t
+                end
+              end
+              types.each do |type|
                 used_variable_in_type(type) do |var|
                   type_params.delete_if { |type_param| type_param.name == var.name }
                 end
